+++ src/module/app.authModule.ts
@@ -3,7 +3,7 @@ import { UserModule } from './app.userModule';
 import { JwtService } from '@nestjs/jwt';
 import { AuthController } from 'src/controller/app.authController';
 import { APP_GUARD } from '@nestjs/core';
-import { AuthGuard } from 'src/guard/app.authGuard';
+import { AuthGuard } from 'src/middleware/app.authGuard';
 import { AuthService } from 'src/service/app.authService';
 
 @Module({
